backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-generativeai
python-dotenv
pydantic
